extract web app for container presenter into explorer common
